Fix warnings: Don't loose integer precision, use proper format specifiers

+++ SPiDSDK/NSData+Base64.m
@@ -104,7 +104,7 @@ static const short _base64DecodingTable[256] = {
     char *objPointer;
     char *strResult;
 
-    int intLength = [self length];
+    NSUInteger intLength = [self length];
     if (intLength == 0) return nil;
 
     strResult = (char *) calloc((size_t) ((((intLength + 2) / 3) * 4) + 1), sizeof(char));
